Rename 'comparator' package to 'comparison' to avoid ambiguity with core java interface
