+++ backend/src/main/java/net/es/oscars/resv/svc/ConnService.java
@@ -9,8 +9,8 @@ import net.es.oscars.app.exc.PCEException;
 import net.es.oscars.app.util.DbAccess;
 import net.es.oscars.app.util.PrettyPrinter;
 import net.es.oscars.model.Interval;
-import net.es.oscars.resv.svc.comparators.ConnServiceBWtoAvailableCompare;
-import net.es.oscars.resv.svc.comparators.ConnServiceVlanToAvailableCompare;
+import net.es.oscars.resv.svc.comparisons.ConnServiceBWtoAvailableCompare;
+import net.es.oscars.resv.svc.comparisons.ConnServiceVlanToAvailableCompare;
 import net.es.oscars.resv.svc.populators.ConnServiceFixtureRequestPopulate;
 import net.es.oscars.resv.svc.populators.ConnServiceFixtureValidityPopulate;
 import net.es.oscars.resv.svc.populators.ConnServicePipeAndEroValidityPopulate;
@@ -24,7 +24,6 @@ import net.es.oscars.sb.nso.resv.NsoResvException;
 import net.es.oscars.resv.db.*;
 import net.es.oscars.resv.ent.*;
 import net.es.oscars.resv.enums.*;
-import net.es.topo.common.model.oscars1.IntRange;
 import net.es.oscars.topo.beans.PortBwVlan;
 import net.es.oscars.web.beans.*;
 import net.es.oscars.web.simple.*;
