--- backend/src/main/java/net/es/oscars/resv/svc/comparators/Comparator.java
@@ -1,6 +1,0 @@
-package net.es.oscars.resv.svc.comparators;
-
-public interface Comparator {
-    void compare();
-    boolean hasErrors();
-}
