+++ backend/src/main/java/net/es/oscars/resv/svc/comparisons/Comparison.java
@@ -1,0 +1,6 @@
+package net.es.oscars.resv.svc.comparisons;
+
+public interface Comparison {
+    void compare();
+    boolean hasErrors();
+}
